+++ src/Hacks/NoRespawnBlink.cpp
@@ -6,18 +6,23 @@
 
 using namespace geode::prelude;
 
-Module* mod = nullptr;
-
 class $modify (CCNode)
 {
     CCAction* runAction(CCAction* action)
     {
-        if (!mod)
-            mod = Client::GetModule("no-blink");
-
-        if (mod->enabled && PlayLayer::get() && (PlayLayer::get()->m_player1 == as<CCNode*>(this) || PlayLayer::get()->m_player2 == as<CCNode*>(this)) && action->getTag() == 11)
+        if (PlayLayer::get() && (PlayLayer::get()->m_player1 == as<CCNode*>(this) || PlayLayer::get()->m_player2 == as<CCNode*>(this)) && action->getTag() == 11)
             return action;
 
         return CCNode::runAction(action);
+    }
+
+    static void onModify(auto& self) {
+        auto hook = self.getHook("CCNode::runAction");
+
+        Loader::get()->queueInMainThread([hook]
+        {
+            auto modu = Client::GetModule("no-blink");
+            modu->addHookRaw(hook);
+        });
     }
 };